Combined module
So now we use PsSec.psm1 as the main module. I made a script called combine in the folder that takes the content of the individual modules and combines them into one file so we can just load that one and not the others. This could just be a task/action, when I push, run this script